+++ R/supplementary_add.R
@@ -1,0 +1,41 @@
+#' Add supplementary data
+#'
+#' Download the employment data for a country and arrange it to the 64x64 SIOTS.
+#' Currently works only with product x product tables. 
+#' @param data_table A SIOT, a use table, a supply table, or a margins table.  
+#' @param supplementary A supplementary account to be added. The key column 
+#' should contain the indicator's name, and the column names must match with the
+#' data_table.
+#' @importFrom dplyr select full_join
+#' @examples
+#' CO2 <- c( 0.2379, 0.5172, 0.0456, 0.1320, 0.0127, 0.0530)  
+#' names ( CO2) <- c("agriculture_group", "industry_group","construction",
+#'                   "trade_group","business_services_group","other_services_group") 
+#' CO2 <- cbind ( 
+#'   tibble ( iotables_row = "CO2"),as.tibble ( t(CO2)))
+#' de_coeff <- input_coefficient_matrix_create ( iotable_get() )
+#'
+#' supplementary_add ( de_io, CO2)
+#'
+#' @export
+
+
+supplementary_add <- function ( data_table, supplementary ) {
+  
+  new_key <- as.character(supplementary[1,1])
+  
+  key_column <- dplyr::select ( data_table, 1 ) 
+  
+  names ( supplementary)[1] <- names (data_table)[1]
+  
+  siot_ext   <- dplyr::full_join ( data_table, supplementary,
+                                   by = names (supplementary) )
+  
+  if ( any(c("final_consumption_households", "p3_s14") %in% tolower ( names ( siot_ext)))  ) {
+    household_col <- which ( tolower ( names ( siot_ext)) %in% c("final_consumption_households", "p3_s14") )
+    new_row <- which ( tolower ( as.character(siot_ext[,1])) %in% tolower(new_key) )
+    siot_ext[new_row, household_col] <- ifelse ( is.na(siot_ext[new_row, household_col]), 0, siot_ext[new_row, household_col])
+  }
+  
+ siot_ext
+}